+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [Version 1.5.0](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.5.0) - Bugfix release - 2026-02-02
+
+- Add blank OAuth SSO preset
+- Add OAuth with secret preset
+
 ## [Version 1.4.2](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.4.2) - Bugfix release - 2026-04-17
 
 - Fix issue with writing single value to static AF attributes
+
 
 ## [Version 1.4.1](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.4.1) - Bugfix release - 2026-01-27